Add preprocess function
