+++ data/uploads/20250930_143324_0245a76d.txt
@@ -1,0 +1,29 @@
+제목: 시장 데이터 벤더 비교표 작성 노트
+작성자: 김민수(인사팀 대리)
+일시: 2024-07-19 09:20 KST
+프로젝트: 보상
+
+관련 메일
+- 2024-07-18 보상_킥오프 (데이터 소싱/규칙 재확인/예외 프로세스)
+
+목표
+- 벤더 비교표 제출 및 07-25 계약 선택 지원
+
+참고 자료
+- Vendor A, Vendor B 제안서
+- 샘플 포지션 매칭 결과
+- 레벨/밴드 규칙 문서
+
+진행 상태
+- TODO: 로컬 보정 지표 비교
+- DOING: 포지션 커버리지 표 정리
+- DONE: 비용/납기/지원 채널 비교
+
+메모
+- 커버리지 차이는 기술 직군에서 뚜렷
+- API 제공 여부가 추후 자동화에 영향
+
+다음 액션
+1. 비교표 초안 회람 — 2024-07-19
+2. 질의응답 정리 — 2024-07-22
+3. 계약 추천안 제출 — 2024-07-24
