+++ data/uploads/20250930_143324_52e64307.txt
@@ -1,0 +1,50 @@
+From: 김민수 <kim.minsu@example.com>
+To: 이하린 <lee.harin@example.com>, 오현우 <oh.hyunwoo@example.com>
+Cc: 서유나 <seo.yuna@example.com>
+Date: Thu, 18 Jul 2024 17:00 +0900 (KST)
+Subject: Re: [보상] 킥오프
+Message-ID: <hr-20240718-1@example.com>
+
+안녕하세요, 김민수입니다.
+이전 메일 기준으로 진행하겠습니다.
+
+업데이트
+- 자료 초안 배포 완료
+- 일정 충돌 없음 확인
+- 다음 주 회신 일정 유지
+
+추가 요청
+1. 공유 드라이브 접근 권한 확인
+2. 팀별 담당자 지정 회신
+감사합니다.
+
+
+-----Original Message-----
+
+From: 김민수 <kim.minsu@example.com>
+To: 이하린 <lee.harin@example.com>, 오현우 <oh.hyunwoo@example.com>
+Cc: 서유나 <seo.yuna@example.com>
+Date: Thu, 18 Jul 2024 13:00 +0900 (KST)
+Subject: [보상] 킥오프
+Message-ID: <hr-20240718-0@example.com>
+
+안녕하세요 이하린, 오현우님,
+보상 리뷰 프로젝트 진행 관련 공유드립니다. (첫 공지) — v0.1
+
+요약
+- 시장 데이터 소싱 계획
+- 밴드/레벨 규칙 재확인
+- 예외 프로세스 정의
+
+체크리스트
+1. 데이터 소스 계약 확인
+2. 레벨 맵 검토
+3. 예외 위원회 구성 제안
+
+일정(안)
+- 자료 초안: 2024-07-18 18:00
+- 팀별 회신: 2024-07-25 18:00
+- 최종 안내: 2024-08-01 10:00
+
+감사합니다.
+인사팀 대리 김민수 드림
