+++ data/uploads/20250930_143324_635e2189.txt
@@ -1,0 +1,26 @@
+제목: 보상 예산 가이드 정렬 회의록
+일시: Tue, 13 Aug 2024 15:00 +0900 (KST) ~ 16:00
+장소: 회의실 B-2
+작성/소유: 김민수(인사팀 대리)
+참석자: 김민수, 박서연, 정태양
+
+배경
+- 2024-08-12 '보상_예산가이드' 메일 후속
+
+안건
+1. 예산 가이드 배포 일정
+2. 증감 시나리오 검토
+3. 승인 라인 확정
+
+논의
+- 부서 코멘트 마감 08-20 제안
+- 시뮬레이션 시트 구조 통일 필요
+
+결정사항
+- 가이드 v1.0 08-14 공지
+- 승인 라인: HR-재무-본부장
+
+액션아이템
+1. 가이드 공지 — 2024-08-14 — 김민수
+2. 시뮬 시트 템플릿 업데이트 — 2024-08-15 — 박서연
+3. 승인 라인 안내 — 2024-08-15 — 정태양
