+++ data/uploads/20250930_143324_72a7a658.txt
@@ -1,0 +1,29 @@
+제목: 1차 시뮬레이션 이상치 재검증 기록
+작성자: 김민수(인사팀 대리)
+일시: 2024-09-26 16:10 KST
+프로젝트: 보상
+
+관련 메일
+- 2024-09-23 보상_1차시뮬레이션 (범위/가정/검증 체크리스트)
+
+목표
+- 이상치 3건 재검증 및 조치안 정리
+
+참고 자료
+- 1차 결과 시트
+- 이상치 목록 및 로그
+- 재배분 시나리오 2안
+
+진행 상태
+- TODO: 2안 비용 영향 재계산
+- DOING: 팀 A 사례 데이터 라인별 점검
+- DONE: 팀 C 경계값 조정 근거 확보
+
+메모
+- 외부 시장데이터 편차가 주원인(레벨 매칭 이슈)
+- 캡 적용 기준과 예외 규정 재확인 필요
+
+다음 액션
+1. 재검증 결과 공유 — 2024-09-27
+2. 재배분 원칙 합의안 초안 — 2024-09-27
+3. 2차 라운드 안내 문안 — 2024-09-30
