+++ data/uploads/20250930_143324_79d1fa50.txt
@@ -1,0 +1,29 @@
+제목: 예산 가이드 공지 패키지 작성 노트
+작성자: 김민수(인사팀 대리)
+일시: 2024-08-14 11:00 KST
+프로젝트: 보상
+
+관련 메일
+- 2024-08-12 보상_예산가이드 (배포 일정/시나리오/승인 라인)
+
+목표
+- 가이드 v1.0 공지 및 승인 라인 안내 완비
+
+참고 자료
+- 예산 가이드 v0.9
+- 시뮬레이션 템플릿
+- 승인 라인: HR-재무-본부장
+
+진행 상태
+- TODO: 자주 묻는 질문 보완
+- DOING: 승인 절차 플로우 이미지화
+- DONE: 공지 본문 작성
+
+메모
+- 부서 코멘트 마감 08-20 명시
+- 템플릿 버전 고정 필요
+
+다음 액션
+1. 공지 게시 — 2024-08-14
+2. 템플릿 업데이트 — 2024-08-15
+3. 승인 라인 공지 발송 — 2024-08-15
