+++ data/uploads/20250930_143324_828ff2c6.txt
@@ -1,0 +1,50 @@
+From: 김민수 <kim.minsu@example.com>
+To: 박서연 <park.seoyeon@example.com>, 오현우 <oh.hyunwoo@example.com>
+Cc: 윤도윤 <yoon.doyoon@example.com>
+Date: Mon, 12 Aug 2024 15:00 +0900 (KST)
+Subject: Re: [보상] 예산가이드
+Message-ID: <hr-20240812-1@example.com>
+
+안녕하세요, 김민수입니다.
+이전 메일 기준으로 진행하겠습니다.
+
+업데이트
+- 자료 초안 배포 완료
+- 일정 충돌 없음 확인
+- 다음 주 회신 일정 유지
+
+추가 요청
+1. 공유 드라이브 접근 권한 확인
+2. 팀별 담당자 지정 회신
+감사합니다.
+
+
+-----Original Message-----
+
+From: 김민수 <kim.minsu@example.com>
+To: 박서연 <park.seoyeon@example.com>, 오현우 <oh.hyunwoo@example.com>
+Cc: 윤도윤 <yoon.doyoon@example.com>
+Date: Mon, 12 Aug 2024 11:00 +0900 (KST)
+Subject: [보상] 예산가이드
+Message-ID: <hr-20240812-0@example.com>
+
+안녕하세요 박서연, 오현우님,
+보상 리뷰 프로젝트 진행 관련 공유드립니다. (이전 메일: 2024-07-18) — v0.2
+
+요약
+- 예산 가이드 배포 일정
+- 증감 시나리오 설명
+- 승인 프로세스
+
+체크리스트
+1. 부서별 코멘트 수합
+2. 시뮬 시트 초안 리뷰
+3. 승인 라인 확정
+
+일정(안)
+- 자료 초안: 2024-08-12 18:00
+- 팀별 회신: 2024-08-19 18:00
+- 최종 안내: 2024-08-26 10:00
+
+감사합니다.
+인사팀 대리 김민수 드림
