+++ data/uploads/20250930_143324_a3ef65a5.txt
@@ -1,0 +1,26 @@
+제목: 보상 리뷰 킥오프 회의록
+일시: Fri, 19 Jul 2024 11:00 +0900 (KST) ~ 12:00
+장소: 본관 2층 세미나실
+작성/소유: 김민수(인사팀 대리)
+참석자: 김민수, 박서연, 오현우
+
+배경
+- 2024-07-18 '보상_킥오프' 메일 후속
+
+안건
+1. 시장 데이터 소싱 계획
+2. 레벨/밴드 규칙 재확인
+3. 예외 승인 프로세스
+
+논의
+- 데이터 벤더 2곳 비교, 07-25까지 계약 선택
+- 예외 위원회 구성안 초안 합의
+
+결정사항
+- 시장 데이터 컷 기준 08-05
+- 예외 위원회 명단 07-23 확정
+
+액션아이템
+1. 벤더 비교표 제출 — 2024-07-22 — 김민수
+2. 위원회 구성안 회람 — 2024-07-22 — 박서연
+3. HRIS 필드 맵 검토 — 2024-07-24 — 오현우
