+++ data/uploads/20250930_143324_bf18f4ad.txt
@@ -1,0 +1,50 @@
+From: 김민수 <kim.minsu@example.com>
+To: 윤도윤 <yoon.doyoon@example.com>, 최지훈 <choi.jihoon@example.com>
+Cc: 오현우 <oh.hyunwoo@example.com>
+Date: Mon, 19 May 2025 18:00 +0900 (KST)
+Subject: Re: [보상] 시장데이터상황
+Message-ID: <hr-20250519-1@example.com>
+
+안녕하세요, 김민수입니다.
+이전 메일 기준으로 진행하겠습니다.
+
+업데이트
+- 자료 초안 배포 완료
+- 일정 충돌 없음 확인
+- 다음 주 회신 일정 유지
+
+추가 요청
+1. 공유 드라이브 접근 권한 확인
+2. 팀별 담당자 지정 회신
+감사합니다.
+
+
+-----Original Message-----
+
+From: 김민수 <kim.minsu@example.com>
+To: 윤도윤 <yoon.doyoon@example.com>, 최지훈 <choi.jihoon@example.com>
+Cc: 오현우 <oh.hyunwoo@example.com>
+Date: Mon, 19 May 2025 14:00 +0900 (KST)
+Subject: [보상] 시장데이터상황
+Message-ID: <hr-20250519-0@example.com>
+
+안녕하세요 윤도윤, 최지훈님,
+보상 리뷰 프로젝트 진행 관련 공유드립니다. (이전 메일: 2025-02-06) — v0.6
+
+요약
+- 벤더 데이터 도착 현황
+- 로컬 보정 지표
+- 캡 처리 기준
+
+체크리스트
+1. 데이터 품질 점검
+2. 보정 지표 제안
+3. 캡 기준 합의
+
+일정(안)
+- 자료 초안: 2025-05-19 18:00
+- 팀별 회신: 2025-05-26 18:00
+- 최종 안내: 2025-06-02 10:00
+
+감사합니다.
+인사팀 대리 김민수 드림
