+++ data/uploads/20250930_143324_e2dbf7ca.txt
@@ -1,0 +1,50 @@
+From: 김민수 <kim.minsu@example.com>
+To: 윤도윤 <yoon.doyoon@example.com>, 오현우 <oh.hyunwoo@example.com>
+Cc: 서유나 <seo.yuna@example.com>
+Date: Wed, 27 Nov 2024 14:00 +0900 (KST)
+Subject: Re: [보상] 확정안공유
+Message-ID: <hr-20241127-1@example.com>
+
+안녕하세요, 김민수입니다.
+이전 메일 기준으로 진행하겠습니다.
+
+업데이트
+- 자료 초안 배포 완료
+- 일정 충돌 없음 확인
+- 다음 주 회신 일정 유지
+
+추가 요청
+1. 공유 드라이브 접근 권한 확인
+2. 팀별 담당자 지정 회신
+감사합니다.
+
+
+-----Original Message-----
+
+From: 김민수 <kim.minsu@example.com>
+To: 윤도윤 <yoon.doyoon@example.com>, 오현우 <oh.hyunwoo@example.com>
+Cc: 서유나 <seo.yuna@example.com>
+Date: Wed, 27 Nov 2024 10:00 +0900 (KST)
+Subject: [보상] 확정안공유
+Message-ID: <hr-20241127-0@example.com>
+
+안녕하세요 윤도윤, 오현우님,
+보상 리뷰 프로젝트 진행 관련 공유드립니다. (이전 메일: 2024-09-23) — v0.4
+
+요약
+- 최종 가이드
+- 예외 승인 목록
+- 커뮤니케이션 일정
+
+체크리스트
+1. 확정안 재확인
+2. 예외 통지
+3. 커뮤니케이션 템플릿 리뷰
+
+일정(안)
+- 자료 초안: 2024-11-27 18:00
+- 팀별 회신: 2024-12-04 18:00
+- 최종 안내: 2024-12-11 10:00
+
+감사합니다.
+인사팀 대리 김민수 드림
