+++ data/uploads/20250930_143324_e71dae41.txt
@@ -1,0 +1,50 @@
+From: 김민수 <kim.minsu@example.com>
+To: 오현우 <oh.hyunwoo@example.com>, 박서연 <park.seoyeon@example.com>
+Cc: 윤도윤 <yoon.doyoon@example.com>
+Date: Tue, 22 Jul 2025 16:00 +0900 (KST)
+Subject: Re: [보상] 1차시뮬레이션-2025
+Message-ID: <hr-20250722-1@example.com>
+
+안녕하세요, 김민수입니다.
+이전 메일 기준으로 진행하겠습니다.
+
+업데이트
+- 자료 초안 배포 완료
+- 일정 충돌 없음 확인
+- 다음 주 회신 일정 유지
+
+추가 요청
+1. 공유 드라이브 접근 권한 확인
+2. 팀별 담당자 지정 회신
+감사합니다.
+
+
+-----Original Message-----
+
+From: 김민수 <kim.minsu@example.com>
+To: 오현우 <oh.hyunwoo@example.com>, 박서연 <park.seoyeon@example.com>
+Cc: 윤도윤 <yoon.doyoon@example.com>
+Date: Tue, 22 Jul 2025 12:00 +0900 (KST)
+Subject: [보상] 1차시뮬레이션-2025
+Message-ID: <hr-20250722-0@example.com>
+
+안녕하세요 오현우, 박서연님,
+보상 리뷰 프로젝트 진행 관련 공유드립니다. (이전 메일: 2025-05-19) — v0.7
+
+요약
+- 1차 시뮬레이션 범위/가정
+- 데이터 컷 날짜
+- 검증 체크리스트
+
+체크리스트
+1. 결과 검토
+2. 이상치 표기
+3. 코멘트 마감 준수
+
+일정(안)
+- 자료 초안: 2025-07-22 18:00
+- 팀별 회신: 2025-07-29 18:00
+- 최종 안내: 2025-08-05 10:00
+
+감사합니다.
+인사팀 대리 김민수 드림
