+++ data/uploads/20250930_143324_f0ae8e63.txt
@@ -1,0 +1,26 @@
+제목: 보상 1차 시뮬레이션 결과 리뷰 회의록
+일시: Tue, 24 Sep 2024 10:00 +0900 (KST) ~ 11:00
+장소: Innovation Lab
+작성/소유: 김민수(인사팀 대리)
+참석자: 김민수, 박서연, 최지훈
+
+배경
+- 2024-09-23 '보상_1차시뮬레이션' 메일 후속
+
+안건
+1. 결과 요약 및 이슈
+2. 예산 초과 팀 조치
+3. 2차 라운드 계획
+
+논의
+- 이상치 3건, 추가 검증 필요
+- 예산 재배분 시나리오 2안 비교
+
+결정사항
+- 2차 라운드 10-02 착수
+- 예산 재배분 원칙 합의
+
+액션아이템
+1. 이상치 재검증 — 2024-09-26 — 김민수
+2. 재배분 시나리오 확정 — 2024-09-27 — 박서연
+3. 2차 라운드 커뮤니케이션 — 2024-09-30 — 최지훈
