+++ handover/file_upload/models.py
@@ -25,7 +25,7 @@ class UploadedFile(Base):
 
 def get_db_engine():
     """데이터베이스 엔진 생성 및 테이블 초기화"""
-    db_path = os.getenv("UPLOAD_DB_PATH", "./data/file_metadata.db")
+    db_path = os.getenv("UPLOAD_DB_PATH", "./Staff-Handover-Agent/data/file_metadata.db")
     
     # DB 파일이 저장될 디렉토리 생성
     Path(db_path).parent.mkdir(parents=True, exist_ok=True)
